// src/app/generated/gtfs-realtime.ts
export interface TripUpdate {
  /**
   * The Trip that this message applies to. There can be at most one
   * TripUpdate entity for each actual trip instance.
   * If there is none, that means there is no prediction information available.
   * It does *not* mean that the trip is progressing according to schedule.
   */
  trip:
    | TripDescriptor
    | undefined;
  /** Additional information on the vehicle that is serving this trip. */
  vehicle?:
    | VehicleDescriptor
    | undefined;
  /**
   * Updates to StopTimes for the trip (both future, i.e., predictions, and in
   * some cases, past ones, i.e., those that already happened).
   * The updates must be sorted by stop_sequence, and apply for all the
   * following stops of the trip up to the next specified one.
   *
   * Example 1:
   * For a trip with 20 stops, a StopTimeUpdate with arrival delay and departure
   * delay of 0 for stop_sequence of the current stop means that the trip is
   * exactly on time.
   *
   * Example 2:
   * For the same trip instance, 3 StopTimeUpdates are provided:
   * - delay of 5 min for stop_sequence 3
   * - delay of 1 min for stop_sequence 8
   * - delay of unspecified duration for stop_sequence 10
   * This will be interpreted as:
   * - stop_sequences 3,4,5,6,7 have delay of 5 min.
   * - stop_sequences 8,9 have delay of 1 min.
   * - stop_sequences 10,... have unknown delay.
   */
  stopTimeUpdate: TripUpdate_StopTimeUpdate[];
  /**
   * The most recent moment at which the vehicle's real-time progress was measured
   * to estimate StopTimes in the future. When StopTimes in the past are provided,
   * arrival/departure times may be earlier than this value. In POSIX
   * time (i.e., the number of seconds since January 1st 1970 00:00:00 UTC).
   */
  timestamp?:
    | number
    | undefined;
  /**
   * The current schedule deviation for the trip.  Delay should only be
   * specified when the prediction is given relative to some existing schedule
   * in GTFS.
   *
   * Delay (in seconds) can be positive (meaning that the vehicle is late) or
   * negative (meaning that the vehicle is ahead of schedule). Delay of 0
   * means that the vehicle is exactly on time.
   *
   * Delay information in StopTimeUpdates take precedent of trip-level delay
   * information, such that trip-level delay is only propagated until the next
   * stop along the trip with a StopTimeUpdate delay value specified.
   *
   * Feed providers are strongly encouraged to provide a TripUpdate.timestamp
   * value indicating when the delay value was last updated, in order to
   * evaluate the freshness of the data.
   *
   * NOTE: This field is still experimental, and subject to change. It may be
   * formally adopted in the future.
   */
  delay?: number | undefined;
  tripProperties?: TripUpdate_TripProperties | undefined;
}
// ...
export interface TripUpdate_StopTimeUpdate {
  /** Must be the same as in stop_times.txt in the corresponding GTFS feed. */
  stopSequence?:
    | number
    | undefined;
  /** Must be the same as in stops.txt in the corresponding GTFS feed. */
  stopId?: string | undefined;
  arrival?: TripUpdate_StopTimeEvent | undefined;
  departure?:
    | TripUpdate_StopTimeEvent
    | undefined;
  /**
   * Expected occupancy after departure from the given stop.
   * Should be provided only for future stops.
   * In order to provide departure_occupancy_status without either arrival or
   * departure StopTimeEvents, ScheduleRelationship should be set to NO_DATA.
   */
  departureOccupancyStatus?: VehiclePosition_OccupancyStatus | undefined;
  scheduleRelationship?:
    | TripUpdate_StopTimeUpdate_ScheduleRelationship
    | undefined;
  /**
   * Realtime updates for certain properties defined within GTFS stop_times.txt
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  stopTimeProperties?: TripUpdate_StopTimeUpdate_StopTimeProperties | undefined;
}
// ...
export interface TripDescriptor {
  /**
   * The trip_id from the GTFS feed that this selector refers to.
   * For non frequency-based trips, this field is enough to uniquely identify
   * the trip. For frequency-based trip, start_time and start_date might also be
   * necessary. When schedule_relationship is DUPLICATED within a TripUpdate, the trip_id identifies the trip from
   * static GTFS to be duplicated. When schedule_relationship is DUPLICATED within a VehiclePosition, the trip_id
   * identifies the new duplicate trip and must contain the value for the corresponding TripUpdate.TripProperties.trip_id.
   */
  tripId?:
    | string
    | undefined;
  /** The route_id from the GTFS that this selector refers to. */
  routeId?:
    | string
    | undefined;
  /**
   * The direction_id from the GTFS feed trips.txt file, indicating the
   * direction of travel for trips this selector refers to.
   */
  directionId?:
    | number
    | undefined;
  /**
   * The initially scheduled start time of this trip instance.
   * When the trip_id corresponds to a non-frequency-based trip, this field
   * should either be omitted or be equal to the value in the GTFS feed. When
   * the trip_id correponds to a frequency-based trip, the start_time must be
   * specified for trip updates and vehicle positions. If the trip corresponds
   * to exact_times=1 GTFS record, then start_time must be some multiple
   * (including zero) of headway_secs later than frequencies.txt start_time for
   * the corresponding time period. If the trip corresponds to exact_times=0,
   * then its start_time may be arbitrary, and is initially expected to be the
   * first departure of the trip. Once established, the start_time of this
   * frequency-based trip should be considered immutable, even if the first
   * departure time changes -- that time change may instead be reflected in a
   * StopTimeUpdate.
   * Format and semantics of the field is same as that of
   * GTFS/frequencies.txt/start_time, e.g., 11:15:35 or 25:15:35.
   */
  startTime?:
    | string
    | undefined;
  /**
   * The scheduled start date of this trip instance.
   * Must be provided to disambiguate trips that are so late as to collide with
   * a scheduled trip on a next day. For example, for a train that departs 8:00
   * and 20:00 every day, and is 12 hours late, there would be two distinct
   * trips on the same time.
   * This field can be provided but is not mandatory for schedules in which such
   * collisions are impossible - for example, a service running on hourly
   * schedule where a vehicle that is one hour late is not considered to be
   * related to schedule anymore.
   * In YYYYMMDD format.
   */
  startDate?: string | undefined;
  scheduleRelationship?:
    | TripDescriptor_ScheduleRelationship
    | undefined;
  /**
   * Linkage to any modifications done to this trip (shape changes, removal or addition of stops).
   * If this field is provided, the `trip_id`, `route_id`, `direction_id`, `start_time`, `start_date` fields of the `TripDescriptor` MUST be left empty, to avoid confusion by consumers that aren't looking for the `ModifiedTripSelector` value.
   */
  modifiedTrip?: TripDescriptor_ModifiedTripSelector | undefined;
}
// ...
export interface TripDescriptor_ModifiedTripSelector {
  /** The 'id' from the FeedEntity in which the contained TripModifications object affects this trip. */
  modificationsId?:
    | string
    | undefined;
  /** The trip_id from the GTFS feed that is modified by the modifications_id */
  affectedTripId?:
    | string
    | undefined;
  /** The initially scheduled start time of this trip instance, applied to the frequency based modified trip. Same definition as start_time in TripDescriptor. */
  startTime?:
    | string
    | undefined;
  /** The start date of this trip instance in YYYYMMDD format, applied to the modified trip. Same definition as start_date in TripDescriptor. */
  startDate?: string | undefined;
}

/** Identification information for the vehicle performing the trip. */
export interface VehicleDescriptor {
  /**
   * Internal system identification of the vehicle. Should be unique per
   * vehicle, and can be used for tracking the vehicle as it proceeds through
   * the system.
   */
  id?:
    | string
    | undefined;
  /**
   * User visible label, i.e., something that must be shown to the passenger to
   * help identify the correct vehicle.
   */
  label?:
    | string
    | undefined;
  /** The license plate of the vehicle. */
  licensePlate?: string | undefined;
  wheelchairAccessible?: VehicleDescriptor_WheelchairAccessible | undefined;
}
// ...
export const VehicleDescriptor: MessageFns<VehicleDescriptor> = {
  encode(message: VehicleDescriptor, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.label !== undefined && message.label !== "") {
      writer.uint32(18).string(message.label);
    }
    if (message.licensePlate !== undefined && message.licensePlate !== "") {
      writer.uint32(26).string(message.licensePlate);
    }
    if (message.wheelchairAccessible !== undefined && message.wheelchairAccessible !== 0) {
      writer.uint32(32).int32(message.wheelchairAccessible);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VehicleDescriptor {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVehicleDescriptor();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.label = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.licensePlate = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.wheelchairAccessible = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VehicleDescriptor {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      label: isSet(object.label) ? globalThis.String(object.label) : "",
      licensePlate: isSet(object.licensePlate) ? globalThis.String(object.licensePlate) : "",
      wheelchairAccessible: isSet(object.wheelchairAccessible)
        ? vehicleDescriptor_WheelchairAccessibleFromJSON(object.wheelchairAccessible)
        : 0,
    };
  },

  toJSON(message: VehicleDescriptor): unknown {
    const obj: any = {};
    if (message.id !== undefined && message.id !== "") {
      obj.id = message.id;
    }
    if (message.label !== undefined && message.label !== "") {
      obj.label = message.label;
    }
    if (message.licensePlate !== undefined && message.licensePlate !== "") {
      obj.licensePlate = message.licensePlate;
    }
    if (message.wheelchairAccessible !== undefined && message.wheelchairAccessible !== 0) {
      obj.wheelchairAccessible = vehicleDescriptor_WheelchairAccessibleToJSON(message.wheelchairAccessible);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VehicleDescriptor>, I>>(base?: I): VehicleDescriptor {
    return VehicleDescriptor.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VehicleDescriptor>, I>>(object: I): VehicleDescriptor {
    const message = createBaseVehicleDescriptor();
    message.id = object.id ?? "";
    message.label = object.label ?? "";
    message.licensePlate = object.licensePlate ?? "";
    message.wheelchairAccessible = object.wheelchairAccessible ?? 0;
    return message;
  },
};